requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
nltk
fastapi
uvicorn
google-cloud-storage
joblib
